refactor(project): move SBOM generator into internal/shop
Address review feedback: keep cmd/project as cobra surface only and put
the shared CycloneDX generation logic in internal/shop so project ci and
project sbom both call shop.WriteProjectSBOM.
